[CHORE] portfolio model 분리
